refactor : Chat 엔티티의 contents 필드를 값 타입으로 구현, Chat 관련 패키지 분리
- 검증 관련 예외를 담당하는 InvalidException 구현
