+++ src/main/java/louie/hanse/shareplate/repository/ChatRepository.java
@@ -1,7 +1,7 @@
 package louie.hanse.shareplate.repository;
 
 import java.util.Optional;
-import louie.hanse.shareplate.domain.Chat;
+import louie.hanse.shareplate.chat.domain.Chat;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 public interface ChatRepository extends JpaRepository<Chat, Long>, CustomChatRepository {
